Fix 2 msvc warnings
`hl2sdk-cs2\public\tier1\keyvalues3.h(340,55): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data`
`hl2sdk-cs2\public\bitmap\imageformat.h(28,14): warning C4091: 'typedef ': ignored on left of 'DXGI_FORMAT' when no variable is declared`

+++ public/tier1/keyvalues3.h
@@ -337,7 +337,7 @@ public:
 		if (!pszString || !pszString[0])
 			return;
 
-		m_nHashCode = MurmurHash2LowerCase(pszString, strlen(pszString), 0x31415926);
+		m_nHashCode = MurmurHash2LowerCase(pszString, (int)strlen(pszString), 0x31415926);
 		m_pszString = pszString;
 
 #if 0
